Refactoring Set2 ASDC - Assignment3
