+++ src/test/java/com/ibatis/common/xml/NodeletUtilsTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2004-2021 the original author or authors.
+ * Copyright 2004-2022 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -37,12 +37,12 @@ class NodeletUtilsTest {
     props.setProperty("boolean4", "def");
     props.setProperty("boolean5", "");
 
-    assertEquals(true, NodeletUtils.getBooleanAttribute(props, "boolean1", false));
-    assertEquals(false, NodeletUtils.getBooleanAttribute(props, "boolean2", true));
-    assertEquals(false, NodeletUtils.getBooleanAttribute(props, "boolean3", false));
-    assertEquals(false, NodeletUtils.getBooleanAttribute(props, "boolean4", true));
-    assertEquals(false, NodeletUtils.getBooleanAttribute(props, "boolean5", true));
-    assertEquals(true, NodeletUtils.getBooleanAttribute(props, "undef", true));
+    assertEquals(true, GetXmlAttribute.getBooleanAttribute(props, "boolean1", false));
+    assertEquals(false, GetXmlAttribute.getBooleanAttribute(props, "boolean2", true));
+    assertEquals(false, GetXmlAttribute.getBooleanAttribute(props, "boolean3", false));
+    assertEquals(false, GetXmlAttribute.getBooleanAttribute(props, "boolean4", true));
+    assertEquals(false, GetXmlAttribute.getBooleanAttribute(props, "boolean5", true));
+    assertEquals(true, GetXmlAttribute.getBooleanAttribute(props, "undef", true));
   }
 
   /**
@@ -58,18 +58,18 @@ class NodeletUtilsTest {
     props.setProperty("int4", "undef");
     props.setProperty("int5", "");
 
-    assertEquals(0, NodeletUtils.getIntAttribute(props, "int1", -10000));
-    assertEquals(1000, NodeletUtils.getIntAttribute(props, "int2", -10000));
-    assertEquals(-200, NodeletUtils.getIntAttribute(props, "int3", -10000));
+    assertEquals(0, GetXmlAttribute.getIntAttribute(props, "int1", -10000));
+    assertEquals(1000, GetXmlAttribute.getIntAttribute(props, "int2", -10000));
+    assertEquals(-200, GetXmlAttribute.getIntAttribute(props, "int3", -10000));
 
     try {
-      assertEquals(-10000, NodeletUtils.getIntAttribute(props, "int4", -10000));
+      assertEquals(-10000, GetXmlAttribute.getIntAttribute(props, "int4", -10000));
       fail("testGetIntAttribute() should have thrown an exception");
     } catch (NumberFormatException ex) {
     }
 
     try {
-      NodeletUtils.getIntAttribute(props, "int5", -10000);
+      GetXmlAttribute.getIntAttribute(props, "int5", -10000);
       fail("testGetIntAttribute() should have thrown an exception");
     } catch (NumberFormatException ex) {
     }
